fixed css code
abstracted components into their individual files, converted css file into css module for localization of css code and readjusted some styles to accomodate the bg image
