--- frontend-client/app/about/about.css
@@ -1,123 +1,0 @@
-.aboutpage {
-    background-image: url("../../public/images/about-background.gif");
-    background-size: cover;
-    background-repeat: no-repeat;
-    width: 100vw;
-    height: 100vh;
-    max-width: 100%;
-}
-
-.about-header {
-    font-weight: bold;
-    font-size: 50px;
-    text-align: center;
-    margin: 20px;
-    color: white;
-}
-
-.header {
-    font-weight: bold;
-    font-size: 40px;
-}
-  
-.creators {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-  
-.creators .card {
-    margin: 20px;
-    padding: 10px;
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 5px 0 rgba(0, 0, 0, 0.19);
-    border-radius: 10px;
-}
-
-.card-title {
-    color: #AF69EF;
-    font-weight: bold;
-}
-  
-.brains-container {
-    text-align: center;
-    margin: 20px;
-}
-  
-.card-text {
-    font-size: 16px;
-    text-align: left;
-}
-  
-.brains-container h2 {
-    color: #A45EE5;
-}
-  
-.tech-stacks {
-    text-align: center;
-    background-color: black;
-}
-  
-.tech-stacks .header {
-    color: rgb(135, 162, 235);
-}
-  
-.tech-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin: 32px;
-}
-  
-#react-logo {
-    animation: spin 5s linear infinite;
-}
-  
-@keyframes spin {
-    0% { transform: rotate(0deg); }
-    100% { transform: rotate(360deg); }
-}
-  
-.logo {
-    width: 100px;
-    height: 100px;
-    margin-left: 10px;
-    margin-right: 10px;
-}
-  
-.mission {
-    display: flex;
-    flex-direction: column;
-    justify-items: center;
-    align-items: center;
-    margin: 50px;
-    color: #7f749c;
-}
-  
-#motivation-pic {
-    margin: 0rem 0rem 1rem 0rem;
-    width: 720px;
-    max-width: 1200px;
-    height: auto;
-}
-  
-.motivation {
-    display: flex;
-    /* flex-direction: column; */
-    /* justify-items: center; */
-    /* align-items: center; */
-    margin: 50px;
-}
-  
-.motivation-text * {
-    word-wrap: normal;
-    color: #7f749c;
-}
-
-.tech-header {
-    font-family: 'Dale Adventure', sans-serif;
-    font-size: 50px;
-}
-
-.motivation-header {
-    color: #ffae42;
-}
